inference_pcl_api/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.115.0
fsspec==2024.9.0
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.25.1
joblib==1.4.2
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.1.0
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.16.1
python-dateutil==2.8.2
python-multipart==0.0.10
pytz==2024.2
requests==2.31.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.38.6
threadpoolctl==3.5.0
timm==1.0.9
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.4
traitlets==5.11.1
triton==3.0.0
typing_extensions==4.9.0
tzdata==2024.2
unicorn==2.1.0
uvicorn==0.30.6
watchdog==5.0.2
wcwidth==0.2.8
wheel==0.43.0
zipp==3.16.2
